pyproject.toml
[project]
name = "literl"
version = "0.1.0"
description = "LiteRL - Simplified PipelineRL"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "vllm>=0.6.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "psutil>=5.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

